Added an expectation about `convert_temp`'s returned class

+++ tests/testthat/test-convert.R
@@ -11,7 +11,7 @@ test_that("Only accepts correct inputs", {
 # Output testing
 test_that("Outputs are correct", {
   ## Class is correct
-  expect_type(convert_temp(value = 15, from = "C", to = "F"), "numeric")
+  expect_type(convert_temp(value = 15, from = "C", to = "F"), "double")
   ## F & C
   expect_equal(convert_temp(value = 15, from = "C", to = "F"), 59)
   expect_equal(convert_temp(value = 59, from = "F", to = "C"), 15)
